More robust readme instructions
Change to make use of the Stash Pull Request branches, instead of the
projectCode/repositoryName, as these are always available in the
destinationRepository (which is same repo being watch by the trigger).
Using the projectCode/repositoryName (which really should be called
sourceRepositoryOwner/sourceRepositoryName for consistency) may fail if
the user does not access permissions to that project (which is often the
default for private forks in Stash).
http://blogs.atlassian.com/2014/08/how-to-fetch-pull-requests/ has
details of these branches.

It also greatly simplifies building merged PR if wanted, as Stash
already provides this as a branch.