Delete old Build Finish Comments
When Pull Request branches, or their target branches, are regularly
updated, it can result in a lot of old Build Finish Comments on the PR,
be they success/failure.
As the only action on PR is merge it, and this only applies to the most
recent HEAD of PR and the target branch, users/maintainers are generally
only interested in the Build state of this, and not all the previous
ones. So deleting the old ones, and this only having the state of the
most recent build cuts down on the noise. It make the PR Builder comment
behave almost like a Build Status as found on Stash Commits.
(aside: it would great if Stash support Build Status for PR, not just
commits).