Add ability to cancel previously queued and running jobs when a PR is updated
When a pr is updated or if you are using `/merge` in the ref spec and
your destination gets updated a new build starts running and there is
really no reason to keep the previously queued items or running jobs
going.

This change gives the option to abort running builds and remove queued
builds for a updated pr. This saves resources and time when things are
updated quickly.